Struct snapbox::Data

pub struct Data { /* private fields */ }

Test fixture, actual output, or expected result

This provides conveniences for tracking the intended format (binary vs text).

Implementations

impl Data

pub fn binary(raw: impl Into<Vec<u8>>) -> Self

Mark the data as binary (no post-processing)

pub fn text(raw: impl Into<String>) -> Self

Mark the data as text (post-processing)

pub fn new() -> Self

Empty test data

pub fn read_from(path: &Path, binary: Option<bool>) -> Result<Self, Error>

Load test data from a file

pub fn write_to(&self, path: &Path) -> Result<(), Error>

Overwrite a snapshot

pub fn replace_lines(
&mut self,
line_nums: Range<usize>,
text: &str
) -> Result<(), Error>

Update an inline snapshot

pub fn map_text(self, op: impl FnOnce(&str) -> String) -> Self

Post-process text

See utils

pub fn try_text(self) -> Self

Convert from binary to text, if possible

This will do extra binary file detection if detect-encoding feature flag is set

pub fn make_text(&mut self) -> Result<(), Utf8Error>

Coerce to a string

Note: this will not do a binary-content check

pub fn into_string(self) -> Result<String, Utf8Error>

Coerce to a string

Note: this will not do a binary-content check

pub fn as_str(&self) -> Option<&str>

Return the underlying str

Note: this will not inspect binary data for being a valid str.

pub fn as_bytes(&self) -> &[u8]